Removes an unnecessary dependency
OAuth isn't used. Well, it is, by Restforce, but it's not a direct
requirement on this project. Also, HipsterHash is only needed for a
test, so it's now a development dependency instead.

I'm considering removing VCR as a development dependency, but there's
always the possibility that one might want to develop a feature against
a live Salesforce instance, so it's handy to have.